+++ example/src/SampleApp/collections/products_collection.tsx
@@ -273,8 +273,7 @@ export const productsCollection = buildCollection<Product>({
                     name: "External id",
                     dataType: "string"
                 }
-            },
-            expanded: true
+            }
         },
         available_locales: {
             name: "Available locales",